+++ client/src/features/list/components/Forms/Plants/AddPlant.tsx
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useField } from "@/hooks/useField";
 import { useShowFormStore } from "@/stores/showFormStore";
-import { createPlant } from "../../../api";
+import { createPlant } from "@/features/list/api";
 import { notify } from "@/utils";
 import { Buttons } from "@/components/Buttons";
 import { Group } from "@/types";
