+++ client/src/features/list/components/Plants/List.tsx
@@ -1,6 +1,6 @@
 import plusSVG from "@/assets/plus.svg";
-import { AddPlant } from "../../../calendar/components/Forms/Plant/AddPlant";
-import { useShowFormStore } from "../../../../stores/showFormStore";
+import { AddPlant } from "../Forms/Plants/AddPlant";
+import { useShowFormStore } from "@/stores/showFormStore";
 import { Group as GroupType, Plant as PlantType } from "@/types";
 import { Entry } from "./Entry";
 import Group from "./Group";
